ffi/util: implement fixed-size non-null pointers
Add an alternative to `NonNull` from the standard library which uses
fixed size addresses. This is useful to implement protocols and
interfaces that use foreign-sized pointers (like the linux ABI). The
NonNull type makes it impossible to model foreign sized pointers with
the same type as native sized ones.

The newly introduced Ptr type works like `NonNull` but has a generic
argument that defines the backing type of the address. This allows
using `NonZeroU*` types to get exactly the same behavior as `NonNull`
but with fixed-size integers over a generic argument.

Signed-off-by: David Rheinsberg <david@readahead.eu>