ffi/util: implement integer abstraction
Provide the explicit `Integer` type, to combine some of our previous
trait efforts into a single type. This will allow us to get integer
types that have a fixed size, alignment, and endianness, all in a single
type, while building on the traits we already have.

By combining everything into an explicit type, rather than trait, we can
get the alignment-setting in a single place, rather than having it
spread everywhere. And, at the same time, we hide endianness in the
types as well.

This paves the way for the `Abi` trait that then exposes all the types
to each ffi module that needs them.

Signed-off-by: David Rheinsberg <david@readahead.eu>