ffi/util: add Abi trait
We want to allow implementing FFI interfaces for non-native formats. To
avoid duplicating the FFI interfaces, or requiring macros everywhere, we
instead provide a generic `Abi` trait, which supplies an FFI interface
with all the Abi-related types and helpers.

For now, the Abi-trait is a simple type-alias collection. But more
features might be added in the future.

Signed-off-by: David Rheinsberg <david@readahead.eu>