ffi/util: drop natural From-impl for Pointer
We ship a From-impl for Pointer which simply wraps `new()`. This is
convenient, but at the same time prevents a lot of more custom
From-impls, because it might produce conflicts when `Address` and
`Target` overlap. Hence, drop the mildly-convenient generic From-impl to
pave the way for more useful specific impls.

Signed-off-by: David Rheinsberg <david@readahead.eu>