lib/osi-derive-proc: add direct-derive proc-macro
Add a new proc-macro crate that implements an alternative to `derive`.

The standard `derive` attribute uses a trivial, but problematic way to
place bounds on generic parameters. It always requires every generic to
implement the trait to derive, even if the target type does not use the
generic as a type.

There is a concept called `perfect derive`, where a derive would place
bounds on each type used in the target type. This provides the required
semantics, but suffers from possible circular trait bounds, and was
thus never an option for the standard library.

The `osi-derive-proc` crate implements such a `perfect derive` attribute
by dispatching any derive to a normal Rust macro.

Signed-off-by: David Rheinsberg <david@readahead.eu>