cargo-osiris/cargo: add extended metadata parser
Add an extended metadata parser which will allow traversing the
dependency-metadata of a target build. This also splits the entire
infrastructure into separate Query, Blob, and Metadata objects, for a
cleaner API and better readability.

The new parser is not used, yet, but will be used to parse java/kotlin
source-dirs in the dependency tree.

Signed-off-by: David Rheinsberg <david@readahead.eu>