osi-demo: add demo application
Add the Osiris Demo Application. This is an application that renders a
simple UI on all supported target platforms to show runtime behavior of
the APIs provided by Osiris.

This is only the start of the application. For now, it provides a simple
GTK4-based UI for linux-fdo platforms. Further platform support is
underway.

Signed-off-by: David Rheinsberg <david@readahead.eu>