ci: map IDs through podman
Most of our container images use UID/GID 1000:1000 to preconfigure a
user account. Unfortunately, GitHub actions use 1001:127 by default.
Hence, we want to remap this user when running containers that care for
UID/GID.

Unfortunately, docker never gained support for it, and upstream
development seems dead. Hence, resort to podman, which properly supports
id-mapped mounts. But since this is a rather new podman feature, we run
podman nested in docker, so we do not have to rely on the host to ship a
recent podman version.

Note that we still rely on recent kernel features to be available for
id-mapped mounts. But those are fortunately updated on GitHub, so we are
good to go.

Signed-off-by: David Rheinsberg <david@readahead.eu>