cargo-osiris/cargo: parse android metadata into sets
To extend the Android metadata to include manifest-files, we have to
retain the package-structure of the metadata, rather than merging
everything upfront. This is, because the Android Gradle plugin requires
each package to be compiled into a separate Android library, since each
can only contain a single Android Manifest. This is unfortunate, but we
can accomodate for it.

The other advantage is that we can better control priorities of
duplicates, since we know exactly how the caller specified the paths.

Signed-off-by: David Rheinsberg <david@readahead.eu>