cargo-osiris: avoid `fresh` as term
Cargo uses `fresh` to indicate that a build was skipped because the
artifacts are still `fresh` (meaning they are newer than the sources).
This is a highly confusing term, since it can also be interpreted as
being updated and thus `fresh`.

Avoid the term entirely and use `new` to indicate that new content has
been written.

Signed-off-by: David Rheinsberg <david@readahead.eu>