lib/osi-apis: add application::Context::locate_data()
Introduce a helper to locate the data directory of an application. This
is meant to mediate between different packaging formats and execution
modes. The idea is to detect how an application is executed, and then
yield a path to either a source checkout (when run through `cargo run`)
or an install path (when run through packaging means).

Signed-off-by: David Rheinsberg <david@readahead.eu>