lib/osiris-*: ensure cross-os builds
Ensure that cross-os builds work fine and disable entire crates. We have
the problem that workspaces cannot select default-members based on the
target properties, but instead always pull everything in.

To allow building the entire workspace, we simply guard os-dependent
crates in their lib.rs. This is a bit of a hack, but at least allows
builds to succeed.

Ideally, in the future, we would allow building these crates even on
foreign architectures. However, this requires to audit the dependencies
and ensure they build properly.

Signed-off-by: David Rheinsberg <david@readahead.eu>