build: reduce default workspace members
Reduce the default workspace members to `osi` and the additional binary
packages from `./lib`. We expect everything else to be pulled in by the
`osi` package, unless it is target specific. In this case, we do not
want it to be part of the default set, since it will likely not build.

Signed-off-by: David Rheinsberg <david@readahead.eu>