build: extend default members
Ensure that we explicitly list all default members in the Cargo manifest
of the workspace. Otherwise, tools like `cargo doc` will not know which
crates are dependencies, and which are part of the operation. Similarly,
`cargo test` would not test the dependencies, even though they are
local.

For now, just rely on all `osi-*` crates to be universal, and the
`osiris-*` crates to be platform-dependent (either for a single
platform, or multiple).

Signed-off-by: David Rheinsberg <david@readahead.eu>