cargo-osiris: support osiris metadata in packages
Support specifying the Osiris metadata in the package metadata of a
Cargo package. The format is analoguous to the old TOML format, but
reformatted as JSON (the TOML->JSON conversion is done by Cargo).

The data is not yet used, but merely parsed by the `cargo` module. It
will replace the old TOML data in a follow up.

Signed-off-by: David Rheinsberg <david@readahead.eu>