analyzer: provide entrypoint in library
Provide the main entrypoint in the analyzer library, to allow builds
with `crate-type` set to `bin`.

Move the old entry-point to a `runner` binary, which allows direct
execution from a checkout via `cargo run`. Unfortunately, Cargo does not
support `cargo run --lib`, even if the library is compiled as binary.

Signed-off-by: David Rheinsberg <david@readahead.eu>